+++ src/config/featureFlags/index.ts
@@ -11,6 +11,8 @@ export const FeatureFlagsSchema = z.object({
 
   create_session: z.boolean().optional(),
   edit_agent: z.boolean().optional(),
+
+  dalle: z.boolean().optional(),
 });
 
 // TypeScript 类型，从 Zod schema 生成
@@ -26,4 +28,21 @@ export const DEFAULT_FEATURE_FLAGS: IFeatureFlags = {
 
   create_session: true,
   edit_agent: true,
+
+  dalle: true,
+};
+
+export const mapFeatureFlagsEnvToState = (config: IFeatureFlags) => {
+  return {
+    enableWebrtc: config.webrtc_sync,
+    isAgentEditable: config.edit_agent,
+
+    showCreateSession: config.create_session,
+    showLLM: config.language_model_settings,
+
+    showOpenAIApiKey: config.openai_api_key,
+    showOpenAIProxyUrl: config.openai_proxy_url,
+
+    showDalle: config.dalle,
+  };
 };
