+++ src/store/featureFlags/selectors.ts
@@ -1,12 +1,5 @@
+import { mapFeatureFlagsEnvToState } from '@/config/featureFlags';
+
 import { FeatureFlagStore } from './store';
 
-export const featureFlagsSelectors = {
-  enableWebrtc: (s: FeatureFlagStore) => s.webrtc_sync,
-  isAgentEditable: (s: FeatureFlagStore) => s.edit_agent,
-
-  showCreateSession: (s: FeatureFlagStore) => s.create_session,
-  showLLM: (s: FeatureFlagStore) => s.language_model_settings,
-
-  showOpenAIApiKey: (s: FeatureFlagStore) => s.openai_api_key,
-  showOpenAIProxyUrl: (s: FeatureFlagStore) => s.openai_proxy_url,
-};
+export const featureFlagsSelectors = (s: FeatureFlagStore) => mapFeatureFlagsEnvToState(s);
